requirements.txt
flask
pytesseract
opencv-python
numpy
Pillow
